# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API Document for WePress",
        "title": "WePress API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/documents/download": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Download Document List",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Download Document List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Documents Per Page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wrapper.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.DownloadDocumentsResponseBody"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    }
                }
            }
        },
        "/documents/download/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Download Document",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Download Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wrapper.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.DownloadDocumentResponseBody"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    }
                }
            }
        },
        "/documents/request-upload": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Generate Presigned URL For Upload Document",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Request Upload Document",
                "parameters": [
                    {
                        "description": "Presigned URL Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UploadRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wrapper.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RequestUploadResponseBody"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    }
                }
            }
        },
        "/documents/upload": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Save Document",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Save Document",
                "parameters": [
                    {
                        "description": "Upload Document Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UploadDocumentRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wrapper.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UploadResponseBody"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/wrapper.FailResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.DownloadDocumentResponseBody": {
            "type": "object",
            "required": [
                "created_at",
                "id",
                "metadata",
                "updated_at",
                "url"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2021-08-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "d2728e88-aef1-4822-976a-63bdca2e89f9"
                },
                "metadata": {
                    "$ref": "#/definitions/dto.MetaDataBody"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-08-01T00:00:00Z"
                },
                "url": {
                    "type": "string",
                    "example": "https://bucket.s3-endpoint/object-key"
                }
            }
        },
        "dto.DownloadDocumentsResponseBody": {
            "type": "object",
            "required": [
                "documents"
            ],
            "properties": {
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DownloadDocumentResponseBody"
                    }
                }
            }
        },
...
        "dto.MetaDataBody": {
            "type": "object",
            "required": [
                "extension",
                "mime_type",
                "name",
                "size"
            ],
            "properties": {
                "extension": {
                    "type": "string",
                    "example": "pdf"
                },
                "mime_type": {
                    "type": "string",
                    "example": "application/pdf"
                },
                "name": {
                    "type": "string",
                    "example": "document"
                },
                "size": {
                    "type": "integer",
                    "example": 10485760
                }
            }
        },
...
        "dto.RequestUploadResponseBody": {
            "type": "object",
            "required": [
                "object_key",
                "url"
            ],
            "properties": {
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "object_key": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "dto.UploadDocumentRequestBody": {
            "type": "object",
            "required": [
                "key",
                "metadata"
            ],
            "properties": {
                "key": {
                    "type": "string",
                    "example": "4b793c1a06ea4ea0a2b019e3c04c3f1d/c211f30fbc56484e83ca9f96afaaeb8b"
                },
                "metadata": {
                    "$ref": "#/definitions/dto.MetaDataBody"
                }
            }
        },
        "dto.UploadRequestBody": {
            "type": "object",
            "required": [
                "size"
            ],
            "properties": {
                "size": {
                    "type": "integer",
                    "example": 10485760
                }
            }
        },
        "dto.UploadResponseBody": {
            "type": "object",
            "required": [
                "id",
                "metadata"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/dto.MetaDataBody"
                }
            }
        },
...
        "wrapper.FailResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Example message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "wrapper.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "metadata": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "AccessToken": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. `Bearer jwt_token_string`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}